feat
added user restriction that only allows users to edit and delete their own items

+++ Item Catalog/app/project.py
@@ -1,10 +1,12 @@
-from flask import Flask, render_template, request, redirect, url_for, flash, jsonify, make_response
+from flask import Flask, render_template, request, redirect, \
+    url_for, flash, jsonify, make_response
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Category, Item
+from database_setup import Base, Category, Item, User
 
 from flask import session as login_session
-import random, string
+import random
+import string
 
 from oauth2client.client import flow_from_clientsecrets
 from oauth2client.client import FlowExchangeError
@@ -13,7 +15,8 @@ import json
 import requests
 
 app = Flask(__name__)
-CLIENT_ID = json.loads(open('client_secrets.json', 'r').read())['web']['client_id']
+CLIENT_ID = json.loads(open('client_secrets.json', 'r')
+                       .read())['web']['client_id']
 APPLICATION_NAME = "Item Catalog App"
 
 # MySQL database information
@@ -24,7 +27,8 @@ APPLICATION_NAME = "Item Catalog App"
 # HOST = "127.0.0.1"
 # PORT = "3306"
 # DATABASE = "item_catalog"
-# DB_URI="{}+{}://{}:{}@{}:{}/{}?charset=utf8".format(DIALCT,DRIVER,USERNAME,PASSWORD,HOST,PORT,DATABASE)
+# DB_URI = "{}+{}://{}:{}@{}:{}/{}?charset=utf8"\
+#     .format(DIALCT, DRIVER, USERNAME, PASSWORD, HOST, PORT, DATABASE)
 # engine = create_engine(DB_URI)
 engine = create_engine('sqlite:///restaurantmenu.db?check_same_thread=False')
 
@@ -33,16 +37,23 @@ Base.metadata.bind = engine
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
-# render login button for google account users
+
 @app.route('/login')
-def showLogin():
-    state = ''.join([random.choice(string.ascii_uppercase + string.digits) for i in range(30)])
+def showlogin():
+    """
+    render login button for google account users
+    """
+    state = ''.join([random.choice(string.ascii_uppercase + string.digits)
+                     for i in range(30)])
     login_session['state'] = state
     return render_template('login.html', STATE=state)
 
-# login with google account
+
 @app.route('/gconnect', methods=['POST'])
 def gconnect():
+    """
+    login with google account
+    """
     # Validate state token
     if request.args.get('state') != login_session['state']:
         response = make_response(json.dumps('Invalid state parameter.'), 401)
@@ -57,17 +68,17 @@ def gconnect():
         oauth_flow.redirect_uri = 'postmessage'
         credentials = oauth_flow.step2_exchange(code)
     except FlowExchangeError:
-        response = make_response(json.dumps('Failed to upgrade the authorization code.'), 401)
+        response = make_response(json.dumps(
+            'Failed to upgrade the authorization code.'), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
 
     # Check that the access token is valid.
     access_token = credentials.access_token
-    print(access_token)
-    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s' % access_token)
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
     h = httplib2.Http()
     response = h.request(url, 'GET')
-    print(response)
     result = json.loads(h.request(url, 'GET')[1])
     # If there was an error in the access token info, abort.
     if result.get('error') is not None:
@@ -94,7 +105,8 @@ def gconnect():
     stored_access_token = login_session.get('access_token')
     stored_gplus_id = login_session.get('gplus_id')
     if stored_access_token is not None and gplus_id == stored_gplus_id:
-        response = make_response(json.dumps('Current user is already connected.'),
+        response = make_response(json.dumps(
+            'Current user is already connected.'),
                                  200)
         response.headers['Content-Type'] = 'application/json'
         return response
@@ -109,10 +121,23 @@ def gconnect():
     answer = requests.get(userinfo_url, params=params)
 
     data = answer.json()
-
+    print(data)
+    login_session['account_id'] = data['id']
     login_session['username'] = data['name']
     login_session['picture'] = data['picture']
     login_session['email'] = data['email']
+
+    user = session.query(User).filter_by(account_id=data['id']).first()
+    if user is None:
+        new_user = User(account_id=data['id'],
+                        email=data['email'],
+                        name=data['name'],
+                        given_name=data['given_name'],
+                        account_link=data['link'],
+                        picture_link=data['picture'],
+                        )
+        session.add(new_user)
+        session.commit()
 
     output = ''
     output += '<h1>Welcome, '
@@ -120,24 +145,30 @@ def gconnect():
     output += '!</h1>'
     output += '<img src="'
     output += login_session['picture']
-    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;' \
+              '-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
     flash("you are now logged in as %s" % login_session['username'])
     print("done!")
     return output
 
-# logout
+
 @app.route('/gdisconnect')
 def gdisconnect():
+    """
+    logout
+    """
     access_token = login_session.get('access_token')
     if access_token is None:
         print('Access Token is None')
-        response = make_response(json.dumps('Current user not connected.'), 401)
+        response = make_response(json.dumps(
+            'Current user not connected.'), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
     print('In gdisconnect access token is %s', access_token)
     print('User name is: ')
     print(login_session['username'])
-    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s'\
+          % login_session['access_token']
     h = httplib2.Http()
     result = h.request(url, 'GET')[0]
     print('result is ')
@@ -154,60 +185,88 @@ def gdisconnect():
         return redirect("/index")
         # return response
     else:
-        response = make_response(json.dumps('Failed to revoke token for given user.'), 400)
+        response = make_response(json.dumps(
+            'Failed to revoke token for given user.'), 400)
         response.headers['Content-Type'] = 'application/json'
         return redirect("/index")
         # return response
 
-# main page
+
 @app.route('/')
 @app.route('/index')
 def index():
+    """
+    main page
+    """
     categories = session.query(Category)
     items = session.query(Item)
     status = 1 if "access_token" in login_session else 0
-    return render_template('index.html', categories=categories, items=items, status=status)
+    return render_template('index.html', categories=categories,
+                           items=items, status=status)
+
 
 @app.route('/item/<int:item_id>/JSON')
 def show_item_JSON(item_id):
-    items = session.query(Item).filter_by(id = item_id)
+    items = session.query(Item).filter_by(id=item_id)
     return jsonify(Item=[i.serialize for i in items])
+
 
 @app.route('/catelog/<string:category_name>/items/JSON')
 def show_category_JSON(category_name):
     categories = session.query(Category)
-    category = session.query(Category).filter_by(name = category_name)
-    items = session.query(Item).filter_by(category_id = category[0].id)
+    category = session.query(Category).filter_by(name=category_name)
+    items = session.query(Item).filter_by(category_id=category[0].id)
     return jsonify(Item=[i.serialize for i in items])
 
-# show all available items in JSON format
+
 @app.route('/catelog/JSON')
 def show_all_JSON():
+    """
+    show all available items in JSON format
+    """
     items = session.query(Item)
     return jsonify(Item=[i.serialize for i in items])
 
 # show all available items for a specific category in JSON format
+
+
 @app.route('/catelog/<string:category_name>/items')
 def show_category(category_name):
     categories = session.query(Category)
-    category = session.query(Category).filter_by(name = category_name)
-    items = session.query(Item).filter_by(category_id = category[0].id)
+    category = session.query(Category).filter_by(name=category_name)
+    items = session.query(Item).filter_by(category_id=category[0].id)
     status = 1 if "access_token" in login_session else 0
-    return render_template('category.html',categories=categories,category=category, items=items, status=status)
+    return render_template('category.html', categories=categories,
+                           category=category, items=items, status=status)
 
-# show item in JSON format
+
 @app.route('/item/<int:item_id>/')
 def show_item(item_id):
-    item = session.query(Item).filter_by(id = item_id)
+    """
+    show item in JSON format
+    :param item_id:
+    """
+    item = session.query(Item).filter_by(id=item_id)
     status = 1 if "access_token" in login_session else 0
-    return render_template('item_description.html',item = item, status=status)
+    user = session.query(User).filter_by(
+        account_id=login_session['account_id']).first()
+    return render_template('item_description.html', item=item,
+                           status=status, user=user)
 
-# edit item information
+
 @app.route('/catelog/<string:item_title>/edit', methods=['GET', 'POST'])
 def edit_item(item_title):
+    """
+    edit item information
+    :param item_title:
+    """
     if 'username' not in login_session:
         return redirect('/login')
     editedItem = session.query(Item).filter_by(title=item_title).one()
+    user = session.query(User).filter_by(
+        account_id=login_session['account_id']).first()
+    if editedItem.user_id != user.id:
+        return redirect('/login')
     categories = session.query(Category)
     status = 1 if "access_token" in login_session else 0
     if request.method == 'POST':
@@ -227,14 +286,23 @@ def edit_item(item_title):
         return redirect(url_for('index'))
     else:
         return render_template(
-            'edit_item.html', item_title = item_title, item = editedItem, categories = categories, status=status)
+            'edit_item.html', item_title=item_title, item=editedItem,
+            categories=categories, status=status)
 
-# delete item from a category
+
 @app.route('/catelog/<string:item_title>/delete', methods=['GET', 'POST'])
 def delete_item(item_title):
+    """
+    delete item from a category
+    :param item_title:
+    """
     if 'username' not in login_session:
         return redirect('/login')
     deletedItem = session.query(Item).filter_by(title=item_title).one()
+    user = session.query(User).filter_by(
+        account_id=login_session['account_id']).first()
+    if deletedItem.user_id != user.id:
+        return redirect('/login')
     status = 1 if "access_token" in login_session else 0
     if request.method == 'POST':
         session.delete(deletedItem)
@@ -246,30 +314,42 @@ def delete_item(item_title):
             'delete_item.html', item=deletedItem, status=status)
     return "delete successful"
 
-# add new item under specific category.
+
 @app.route('/catelog/new_item', methods=['GET', 'POST'])
 def new_item():
+    """
+    add new item under specific category.
+    """
     if 'username' not in login_session:
         return redirect('/login')
     categories = session.query(Category)
     status = 1 if "access_token" in login_session else 0
+    user = session.query(User).filter_by(
+        account_id=login_session['account_id']).first()
     if request.method == 'POST':
-        new_item = Item(title = request.form['title'], description = request.form['description'], category_id = request.form['category'])
+        new_item = Item(title=request.form['title'],
+                        description=request.form['description'],
+                        category_id=request.form['category'],
+                        user_id=user.id)
         session.add(new_item)
         session.commit()
         flash("new item created!")
         return redirect(url_for('index'))
     else:
-        return render_template('new_item.html', categories=categories, status=status)
+        return render_template('new_item.html',
+                               categories=categories, status=status)
 
-# add new category.
+
 @app.route('/catelog/new_category', methods=['GET', 'POST'])
 def new_category():
+    """
+    add new category.
+    """
     if 'username' not in login_session:
         return redirect('/login')
     status = 1 if "access_token" in login_session else 0
     if request.method == 'POST':
-        new_category = Category(name = request.form['name'])
+        new_category = Category(name=request.form['name'])
         session.add(new_category)
         session.commit()
         flash("new category created!")
@@ -277,7 +357,8 @@ def new_category():
     else:
         return render_template('new_category.html', status=status)
 
+
 if __name__ == '__main__':
     app.secret_key = "secret_key"
     app.debug = True
-    app.run(host = '0.0.0.0', port = 5000)
+    app.run(host='0.0.0.0', port=5000)
